requirements.txt
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
google-api-core==2.19.1
google-api-python-client==2.140.0
google-auth==2.33.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.63.2
gspread==6.1.2
httplib2==0.22.0
idna==3.7
numpy==2.0.1
oauthlib==3.2.2
pandas==2.2.2
proto-plus==1.24.0
protobuf==5.27.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
